+++ src/lib/darknode/utils/blockStateUtils.ts
@@ -1,17 +1,20 @@
 import { queryBlockStateResponseMock } from "./currentMock";
 
 export type QueryBlockStateResponse = typeof queryBlockStateResponseMock;
-export type QueryBlockState = QueryBlockStateResponse["result"]["state"]["v"];
+export type BlockState = QueryBlockStateResponse["result"]["state"]["v"];
 export type Numeric = number | string;
-export const getCurrentEpochId = (blockState: QueryBlockState) => {
+
+export const getCurrentEpochId = (blockState: BlockState) => {
     return Number(blockState.System.epoch.number);
 };
+
 export const toNativeTokenSymbol = (symbol: string) => {
     return symbol.replace(/^ren/, "").replace(/^test/, "").replace(/^dev/, "");
 };
+
 export const getNodeEnteredAt = (
     renVmNodeId: string,
-    blockState: QueryBlockState,
+    blockState: BlockState,
 ) => {
     const nodeSystemData = blockState.System.nodes.find(
         (node) => node.id === renVmNodeId,
@@ -20,4 +23,8 @@ export const getNodeEnteredAt = (
         return null;
     }
     return Number(nodeSystemData.enteredAt);
+};
+
+export const getNodeExists = (renVmNodeId: string, blockState: BlockState) => {
+    return getNodeEnteredAt(renVmNodeId, blockState) !== null;
 };
